task 7

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  * Discription: prints the last digit of a number
  * Return: 0
  */
-int print_last_digit(int r)
+int print_last_digit(long int r)
 {
 	int i;
